+++ SourceCode/Services/Implementations/LayoutParticipantService.cs
@@ -12,7 +12,7 @@
 
                 return await dbContect.LayoutParticipants.AsNoTracking()
                     .Include(x => x.Layout).ThenInclude(x => x.Meeting)
-                    .Include(x => x.LayoutModules).ThenInclude(x => x.Module)
+                    .Include(x => x.LayoutModules).ThenInclude(x => x.Module).ThenInclude(x => x.ModuleOwnerships)
                     .Include(x => x.Person)
                     .Include(x => x.LayoutStations).ThenInclude(ls => ls.Station)
                     .Where(x => x.LayoutId == layoutId && x.MeetingParticipant.CancellationTime == null)