+++ RELEASENOTES.md
@@ -2,6 +2,13 @@
 > Release notes are only published in English.
 The release notes are summaries of important changes and fixes in each release. 
 There is maximum one release per date, but this can be performed as one or several deployments during a day.
+## Release 1.10.4
+Release date 2024-05-07
+- **Bug fix** of deleting module, that is part of an operational place, from meeting layout. 
+Deleting a module that is part of an operational place (station) will now delete all modules within the same operational places as well as the operational places.
+- **Fix** of issue #198. When a partly owned module is registerd on a meeting layout, the module will be visible
+for all owners that are registered, not just the owner that registered the module to the layout.
+
 ## Release 1.10.3
 Release date 2024-04-24
 - **Bug fix** of mail button emaal link.